binary: Introduce the Header type
The Header struct defines the first 32-bytes of any given moss package.
A number of validation checks are possible with this approach as we
not only check the magic header is correct (NUL MOS), but we have
a fixed integrity check that must pass (hardcoded values) to detect
early stream corruption.

Additionally we define a file **type**, which is currently just binary
or delta. Lastly, version number and record count is set, which will
allow us to scan all records ahead of time, knowing the payload will
exist after the records.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>