store: Download + link into place via sharing
Sharing is currently preferred through a hardlink but may not be
possible when cross-device links are in place, so copies may have
to happen.

In most local cases we'll get a hardlink and all will work well.
This approach ensures each build has a local view of sources that
has been appropriately named files, to avoid system wide cache
clashes.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>