db/state/meta: Add some initial skeleton work on persistence
We have a few placeholders in right now where value encoding/decoding
will eventually be present. However, for the moment we can store a State
into the ECS, locate it and encode to the DB. Remaining items include
cleaning up the implementation and adding some nice accessor APIs as
well as correctly encoding the value, but its a huge boost/step in the
right direction and will serve as a template for our other DB
implementations.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>