moss/db: Set up a new API contract for iteration
We'll rely on wrapping up heavyweight objects in our own GC managed
classes to add `foreach` semantics to property iterators on the database
and buckets.

This way we can nicely hide the details from the consumer and have bucket
based iteration "just work" with proper encapsulation of the key, prefix
and value properties by way of a tuple.

Eventually we'd like to have handle pools to reuse old iterators that
have been left around.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>