storage/db: Incorporate wildly new StateDB
The StateDB is responsible for storing all system "states" within the
database. A state is simply a representation of the full system (OS) at
a given point of time (transaction construction).

Along with some metadata, all current package selections are stored
so that we can remember what was included within each installation.
This will allow persistence of old selections to continue within each
newly constructed rootfs.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>