db/installdb: Start work on storing payloads for precache archives
We now have to query the InstallDB to retrieve the package ID for
a locally installable archive. We gain this by reading the MetaPayload
and begin storing a refCount of 0 for new packages. When we have
some delete logic in place, we'll bump refcount to 1 for all packages
successfully incorporated, and wipe all assets with 0 refcount as
part of a GC sweep.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>