db/metadata: Begin work on skeletal MetadataDB
We'll use the MetadataDB to absorb commonality between our current
InstallDB and the upcoming RepoDB. It will be responsible for handling
queries in relation to providers + dependencies and largely conform
to the moss-deps RegistryPlugin *concepts* but not the API itself.

This will allow replacement plugins to be built based around this
DB while providing efficient filtering as opposed to mutating
various allocated arrays.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>