controller/rootconstructor: Optimize buildPath usage by reusing rootfsDir
As we blit a full root with every change, the use of buildPath was using ~3% of
cycles installing a package, when most of the time is spent extracting the
package! As roots grow bigger and bigger, the time will increase.

Initial testing shows that reducing the number of dirs passed reduces time
needed. Here we reuse the constant rootfsDir which gives us the store root
instead of recalculating it for every single file. This reduces the time
spent in buildPath functions by 14% (source: callgrind) with further
improvements possible.

Signed-off-by: Peter O'Connor <sunnyflunk@serpentos.com>