systemcache: Far more robust caching approach
Previously we used a tmpfs based approach which had the obvious limitation
that on lower-memory devices, we'd very quickly eat all available memory
via decompression of the content payload. Instead we now take a more optimal
approach of decompressing once to the disk, and using the Linux specific
copy_file_range call to blit the indices to disk.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>