moss: Add inspect command w/@safe signature
The actual implementation is mostly ported verbatim from the old
rocksdb-based moss implementation and is marked @trusted and thus
not likely to be @safe to any appreciable degree.

This commit also tweaks the format of formattedSize toString output for
column view purposes. Example output is included in the docstring header.

Signed-off-by: Rune Morling <ermo@serpentos.com>