client: Atomically update real `/usr` tree, not a symlink
We've always used a symlink for `/usr` to the real staging root.
With this change `/usr` is a real tree that we atomically swap with
a staged transaction, before archiving the old transaction into a
resting store.

This now give us a **real** `/usr` directory and will unclobber our
nasty path system!

Signed-off-by: Ikey Doherty <ikey@serpentos.com>